requirements.txt
# Core
faster-whisper>=1.0.0
PyAudio>=0.2.14
httpx>=0.27.0
python-dotenv>=1.0.0
numpy>=1.26.0

# TTS — install the engine(s) you plan to use
#piper-tts>=1.2.0
kokoro>=0.8.0

# Fake hub server (for development without the real learning-hub)
starlette>=0.40.0
uvicorn>=0.30.0
